+++ signup.py
@@ -1,3 +1,18 @@
+def login(cur):
+    username = input("Username: ")
+    password = input("Password: ")
+    select = f"select username, password from users where username = '{username}' and password = '{password}';"
+    cur.execute(select)
+    login = cur.fetchall()
+    # cur.close()
+    if (len(login) == 0):
+        print("Wrong Username/password or Please signup first!!")
+    else :
+        print("Login Successfully!!!")
+        return True
+        
+
+
 def signup(cur):
     while True:
         first_name = input("Please enter First name: ")
@@ -11,9 +26,11 @@ def signup(cur):
         if (len(usernames)==0):
             insert = f"insert into users (first_name,last_name,password,username) values('{first_name}','{last_name}','{password}','{username}');"
             cur.execute(insert)
-            # conn.commit()
-            return ("Signup Successfully")
-            
+            print("Signup Successfully")
+            break
         else :
-             print("Username Already exists, please login!!!")
-             signup.login(cur)
+            print("Username Already exists, please try again!!!")
+             
+
+
+
